Create StringHelper class
The StringHelper class contains methods that swap one representation of
a letter or character for another. This class was created because some
of the names scraped from the html contain ugly alternatives to some of
the characters. Some of the characters have two different versions such
as the dash. These characters were messing up some of the regular
expressions or were not human friendly, so they are replaced.

Note: I think these weird characters are caused by character encoding
and the internet and such. I don't completely understand it, but it's
easy just to replace the weird character representations with the proper
character.