Fix urls not being removed from UrlQueue
The urls to be crawled are stored in an ArrayDequeue. To remove items,
the function removeFirst() was being used. This function only returns
the first item in the ArrayDequeue and does not remove it. The function
was switched to poll() which does remove the item from the ArrayDequeue
before returning it.