Change WebCrawler class to use bridge pattern
Changed the WebCrawler class to use a bridge pattern instead of
inheritance (At least I think I used something that resembles a bridge
pattern...). The WebCrawler class name was changed to
WebCrawlerController. The new abstract WebCrawler class is basically the
public api for the web crawler. This change allowed the links filtering
process to be more efficient. The code (at least I think) is cleaner
with the actual api seperate from most of the core logic.