Add Javadoc and refactor the CrawlerConfig class
Added Javadoc to the CrawlerConfig class. Also changed how seed urls are
added to the configurartion. The code was simplified down to a method
that only takes in one url at a time, instead of mulitple urls. Another
method was added to remove seed urls from the configuration. Some of the
code in the project's main method had to be changed due to the
CrawlerConfig class changing.